+++ src/GameObjects/Player.java
@@ -51,11 +51,11 @@ public class Player extends GameObject {
         Graphics2D graphics2D=(Graphics2D)g;
 
         g.setColor(Color.white);
-        g.fillRect(x,y,32,32);
+        g.fillRect((int)x,(int)y,32,32);
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x,y,32,32);
+        return new Rectangle((int)x,(int)y,32,32);
     }
 }
