+++ src/main/HUD.java
@@ -4,8 +4,8 @@ import java.awt.*;
 
 public class HUD {
 
-    public static int HEALTH=100;
-    private int greenValue=255;
+    public static float HEALTH=100;
+    private float greenValue=255;
 
     private int score=0;
     private int level=1;
@@ -20,8 +20,8 @@ public class HUD {
     }
 
     public void render(Graphics g){
-        g.setColor(new Color(100,greenValue,0));
-        g.fillRect(10,10,HEALTH,5);
+        g.setColor(new Color(100,(int)greenValue,0));
+        g.fillRect(10,10,(int)HEALTH,5);
 
         g.setColor(new Color(255,255,153));
         g.drawString("Score: "+score,540,13);
